refactored all files + fixed bSpotted filter

+++ CS2_Internal_Trainer/GameChecker.h
@@ -4,7 +4,7 @@
 namespace GameChecker
 {
 	static bool bGamePointerInit{false};
-	static constexpr int_least8_t inGameId{ 8 };
-	static constexpr int_least8_t notInGameId{ 4 };
-	inline int_least8_t oldGameStateId{ NULL };
+	static constexpr int16_t inGameId{ 8 };
+	static constexpr int16_t notInGameId{ 4 };
+	inline int16_t oldGameStateId{ NULL };
 };
